Remove now-unused crypto/ec and crypto/bn stuff.
