+++ crypto/bn/bn.c
@@ -212,38 +212,8 @@ int BN_set_word(BIGNUM *bn, BN_ULONG value) {
   return 1;
 }
 
-int bn_get_words(BN_ULONG *words, const BIGNUM *bn, size_t num) {
-  if ((size_t)bn->top > num) {
-    return 0;
-  }
-  for (size_t i = 0; i < num; ++i) {
-    words[i] = (i < (size_t)bn->top) ? bn->d[i] : 0;
-  }
-  return 1;
-}
-
-int bn_set_words(BIGNUM *bn, const BN_ULONG *words, size_t num) {
-  if (bn_wexpand(bn, num) == NULL) {
-    return 0;
-  }
-  memmove(bn->d, words, num * sizeof(BN_ULONG));
-  /* |bn_wexpand| verified that |num| isn't too large. */
-  bn->top = (int)num;
-  bn_correct_top(bn);
-  bn->neg = 0;
-  return 1;
-}
-
 int BN_is_negative(const BIGNUM *bn) {
   return bn->neg != 0;
-}
-
-void BN_set_negative(BIGNUM *bn, int sign) {
-  if (sign && !BN_is_zero(bn)) {
-    bn->neg = 1;
-  } else {
-    bn->neg = 0;
-  }
 }
 
 BIGNUM *bn_wexpand(BIGNUM *bn, size_t words) {
