+++ crypto/libring.Windows.vcxproj
@@ -34,16 +34,11 @@
     <ClCompile Include="bn\shift.c" />
     <ClCompile Include="cipher\e_aes.c" />
     <ClCompile Include="curve25519\curve25519.c" />
-    <ClCompile Include="ec\ec.c" />
-    <ClCompile Include="ec\ec_curves.c" />
-    <ClCompile Include="ec\ec_montgomery.c" />
     <ClCompile Include="ec\ecp_nistz.c" />
     <ClCompile Include="ec\ecp_nistz256.c" />
     <ClCompile Include="ec\gfp_constant_time.c" />
     <ClCompile Include="ec\gfp_p256.c" />
     <ClCompile Include="ec\gfp_p384.c" />
-    <ClCompile Include="ec\gfp_suite_b_public.c" />
-    <ClCompile Include="ec\simple.c" />
     <ClCompile Include="ec\wnaf.c" />
     <ClCompile Include="mem.c" />
     <ClCompile Include="modes\gcm.c" />
