+++ src/ec/suite_b/ops/p256.rs
@@ -56,8 +56,6 @@ pub static COMMON_OPS: CommonOps = CommonOps {
     elem_sqr_mont: ecp_nistz256_sqr_mont,
 
     point_add_jacobian_impl: ecp_nistz256_point_add,
-
-    ec_group: &EC_GROUP_P256,
 };
 
 
@@ -302,8 +300,6 @@ extern {
     fn GFp_p256_scalar_sqr_rep_mont(r: *mut Limb/*[COMMON_OPS.num_limbs]*/,
                                     a: *const Limb/*[COMMON_OPS.num_limbs]*/,
                                     rep: c::int);
-
-    static EC_GROUP_P256: EC_GROUP;
 }
